+++ mycodo/mycodo_flask/routes_page.py
@@ -22,6 +22,7 @@ from flask_babel import gettext
 
 from mycodo.config import ALEMBIC_VERSION
 from mycodo.config import BACKUP_LOG_FILE
+from mycodo.config import CONDITIONALS
 from mycodo.config import CONDITIONAL_ACTIONS
 from mycodo.config import DAEMON_LOG_FILE
 from mycodo.config import DAEMON_PID_FILE
@@ -1043,6 +1044,7 @@ def page_function():
                            choices_input=choices_input,
                            choices_math=choices_math,
                            choices_pid=choices_pid,
+                           conditional_names=CONDITIONALS,
                            conditional=conditional,
                            conditional_actions=conditional_actions,
                            conditional_actions_list=CONDITIONAL_ACTIONS,